qdrant/requirements.txt
google-generativeai
qdrant-client
python-dotenv
numpy